Refactor shared Python tooling
