+++ scripts/lidarslam_tools/__init__.py
@@ -1,0 +1,5 @@
+"""Shared, dependency-light building blocks for repository command-line tools."""
+
+from .serialization import payload_to_json
+
+__all__ = ['payload_to_json']
